+++ src/components/SideNav.vue
@@ -22,6 +22,14 @@
             <v-list-item-title>Starkiller</v-list-item-title>
             <v-list-item-subtitle>{{ version }}</v-list-item-subtitle>
           </v-list-item-content>
+          <v-list-item-icon>
+            <v-btn
+              icon
+              @click.stop="toggle"
+            >
+              <v-icon>mdi-page-last</v-icon>
+            </v-btn>
+          </v-list-item-icon>
         </v-list-item>
 
         <v-divider />
@@ -77,14 +85,14 @@ export default {
         },
         { title: 'Agents', pathName: 'agents', icon: 'fa-link' },
         { title: 'Modules', pathName: 'modules', icon: 'fa-grip-horizontal' },
+        { title: 'Obfuscation', pathName: 'obfuscation', icon: 'fa-magic' },
         { title: 'Credentials', pathName: 'credentials', icon: 'fa-key' },
         { title: 'Reporting', pathName: 'reporting', icon: 'fa-sticky-note' },
-        { title: 'Users', pathName: 'users', icon: 'fa-user' },
-        { title: 'Plugins', pathName: 'plugins', icon: 'fa-plug' },
-        { title: 'Downloads', pathName: 'downloads', icon: 'fa-download' },
-        { title: 'Bypasses', pathName: 'bypasses', icon: 'fa-shield-virus' },
         { title: 'Malleable Profiles', pathName: 'malleableProfiles', icon: 'fa-random' },
         { title: 'Bypasses', pathName: 'bypasses', icon: 'fa-shield-virus' },
+        { title: 'Plugins', pathName: 'plugins', icon: 'fa-plug' },
+        { title: 'Downloads', pathName: 'downloads', icon: 'fa-download' },
+        { title: 'Users', pathName: 'users', icon: 'fa-user' },
         { title: 'Settings', pathName: 'settings', icon: 'fa-cog' },
         { title: 'About', pathName: 'about', icon: 'fa-info-circle' },
       ],